refactor: move Pagination into SharedComponents
