+++ src/components/SharedComponents/Pagination/Pagination.tsx
@@ -2,7 +2,7 @@ import Router, { useRouter } from "next/router";
 import { resolveHref } from "next/dist/shared/lib/router/utils/resolve-href";
 import React, { FC, RefObject, useEffect } from "react";
 
-import { Span } from "../Typography";
+import { Span } from "../../Typography";
 
 import IconButtonAsLink from "@/components/SharedComponents/Button/IconButtonAsLink";
 import Flex from "@/components/SharedComponents/Flex";
