+++ src/components/pages/WebinarsIndex.page.tsx
@@ -10,7 +10,7 @@ import { WebinarsListingPage } from "../../common-lib/cms-types/webinarsListingP
 import PostListing from "../Posts/PostListing";
 import { getVideoThumbnail } from "../VideoPlayer/getVideoThumbnail";
 import { PAGE_SIZE } from "../Posts/PostList/usePostList";
-import usePagination from "../Pagination/usePagination";
+import usePagination from "../SharedComponents/Pagination/usePagination";
 
 export type SerializedWebinarPreview = Omit<WebinarPreview, "date"> & {
   date: string;
