+++ src/pages/teachers/programmes/[programmeSlug]/units/[unitSlug]/lessons.tsx
@@ -10,7 +10,7 @@ import AppLayout from "@/components/AppLayout";
 import { getSeoProps } from "@/browser-lib/seo/getSeoProps";
 import MaxWidth from "@/components/SharedComponents/MaxWidth";
 import LessonList from "@/components/UnitAndLessonLists/LessonList";
-import usePagination from "@/components/Pagination/usePagination";
+import usePagination from "@/components/SharedComponents/Pagination/usePagination";
 import Grid, { GridArea } from "@/components/Grid";
 import {
   getFallbackBlockingConfig,
